Refactor: change package name to allow merging
