+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.WellFed"
         tools:targetApi="31">
         <activity
-            android:name=".Recipe.RecipeActivity"
+            android:name=".Recipe2.RecipeActivity"
             android:exported="false"
             android:parentActivityName=".MainActivity">
             <meta-data
