--- app/src/main/java/com/example/wellfed/Recipe/RecipeController.java
@@ -1,4 +1,0 @@
-package com.example.wellfed.Recipe;
-
-public class RecipeController {
-}
