+++ app/src/main/java/com/example/wellfed/Recipe2/RecipeBookFragment.java
@@ -1,4 +1,4 @@
-package com.example.wellfed.Recipe;
+package com.example.wellfed.Recipe2;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,7 +6,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
