+++ app/src/main/java/com/example/wellfed/Recipe2/RecipeController.java
@@ -1,0 +1,4 @@
+package com.example.wellfed.Recipe2;
+
+public class RecipeController {
+}
