+++ app/src/main/java/com/example/wellfed/Recipe2/RecipeDB.java
@@ -1,6 +1,4 @@
-package com.example.wellfed.Recipe;
-
-import com.example.wellfed.R;
+package com.example.wellfed.Recipe2;
 
 import java.util.ArrayList;
 import java.util.List;
