+++ app/src/main/res/layout/activity_recipe.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Recipe.RecipeActivity">
+    tools:context=".Recipe2.RecipeActivity">
 
     <ImageView
         android:id="@+id/recipe_img"
